move GovernaceData to own module
