+++ toolkit/offchain/src/d_param/tests.rs
@@ -1,5 +1,5 @@
 use super::{mint_d_param_token_tx, update_d_param_tx};
-use crate::init_governance::GovernanceData;
+use crate::governance::GovernanceData;
 use crate::{
 	csl::{empty_asset_name, TransactionContext},
 	test_values::*,
